requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
python-dotenv
Authlib
requests
whisper
Flask-Migrate